+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.5.5
 * Update dependencies
 * Added pre-commit tests
+* Use the first locale if it is absent in URL hash
 
 ## 0.5.4
 * Update dependencies